Page: Rename folder name quantity to amount
