+++ src/app/recommend/explore/(attributes)/amount/page.module.scss
